Adds Aggregator Module
