+++ lib/shifter.rb
@@ -1,5 +1,4 @@
 
-
 module Shifter
 
   def set_keys(key = randomkey)
